requirements.txt
ads==0.12.3
attrs==21.4.0
cached-property==1.5.2
cachetools==4.2.4
certifi==2021.10.8
charset-normalizer==2.0.12
google-ads==15.1.1
google-api-core==2.7.1
google-auth==1.35.0
google-auth-oauthlib==0.5.1
googleapis-common-protos==1.56.0
grpcio==1.44.0
grpcio-status==1.44.0
httpretty==0.8.10
idna==3.3
isodate==0.6.1
lxml==4.8.0
mock==4.0.3
oauthlib==3.2.0
platformdirs==2.5.1
proto-plus==1.19.6
protobuf==3.20.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
python-dotenv==0.20.0
pytz==2022.1
PyYAML==5.4.1
requests==2.27.1
requests-file==1.5.1
requests-oauthlib==1.3.1
requests-toolbelt==0.9.1
rsa==4.8
six==1.16.0
urllib3==1.26.9
Werkzeug==2.1.1
xmltodict==0.12.0
zeep==4.1.0
